Break out related functionalities from Enigma class into Crackable and Generable modules
